// types/bot.ts
export type BotMenuConfig = {
  id: number;
  userId: number;
  menuText: string;
  menuFooterText: string | null;
  menuButtonBuyText: string;
  menuButtonAddBalanceText: string;
  menuButtonSupportText: string;
  categoryListHeaderText: string;
  categoryListBodyText: string;
  categoryListFooterText: string;
  categoryListFooterMoreText: string;
  categoryListButtonText: string;
  categoryListSectionTitle: string;
  categoryListNextTitle: string;
  categoryListNextDescription: string;
  categoryListEmptyText: string;
  categoryDetailBodyText: string;
  categoryDetailFooterText: string;
  categoryDetailButtonText: string;
  categoryDetailFileCaption: string | null;
  addBalanceReplyText: string;
  supportReplyText: string;
  variables: string[];
  imagePath: string | null;
  updatedAt: string;
};